+++ src/main/kotlin/calculator/delimiter/DelimiterSplitter.kt
@@ -1,0 +1,7 @@
+package calculator.delimiter
+
+interface DelimiterSplitter {
+    fun isSupport(text: String): Boolean
+
+    fun split(text: String): List<String>
+}
